Revert "Integrate common methods into parent classes"
