+++ README.md
@@ -144,27 +144,3 @@ Update version info
 - Dockerfile
 - src/my_portscanner/version.py
 - ./pyproject.toml
-
-### in the Devcontainer
-
-#### add `rye` path
-
-In DevContainer, I use the .bashrc of my [dotfiles](https://github.com/RyosukeDTomita/dotfiles/blob/main/.bashrc), which allows the path of rye installed with aqua to pass through.
-DevContainer内では自分のdotfilesの.bashrcを使用しており，これによってaquaでインストールしたryeのpathが通るようにできている。
-
-```shell
-source ~/.bashrc
-```
-OR
-
-```shell
-export PATH="$(aqua root-dir)/bin:$PATH"
-```
-
-#### use black formatter
-
-- To use black formatter VSCode Extensions, you need to activate venv and pass through python.
-
-```shell
-source ~/.venv/bin/activate
-```
